Initial version for flutter_tree.
PiperOrigin-RevId: 312162699
